Split factory implementations for room M370 and the incubator.
