+++ SmartNode/SmartNode/Program.cs
@@ -76,7 +76,16 @@ namespace SmartNode
             // Register a factory to allow for dynamic constructor argument passing through DI.
             builder.Services.AddSingleton<IMongoClient, MongoClient>(serviceProvider => new MongoClient(databaseSettings!.ConnectionString));
             builder.Services.AddSingleton<ICaseRepository, CaseRepository>(serviceProvider => new CaseRepository(serviceProvider));
-            builder.Services.AddSingleton<IFactory, Factory>(serviceProvider => new Factory(serviceProvider, coordinatorSettings!.Environment));
+            builder.Services.AddSingleton<IFactory>(serviceProvider => {
+                // TODO: if we have many more of these, it would be nicer putting them in a dictionary.
+                if (coordinatorSettings!.Environment.Equals("roomM370")) {
+                    return new RoomM370Factory(serviceProvider);
+                } else if (coordinatorSettings.Environment.Equals("incubator")) {
+                    return new IncubatorFactory(serviceProvider);
+                } else {
+                    throw new Exception($"No factory found for environment {coordinatorSettings.Environment}.");
+                }
+            });
             builder.Services.AddSingleton<IMapekMonitor, MapekMonitor>(serviceProvider => new MapekMonitor(serviceProvider));
             builder.Services.AddSingleton<IMapekPlan, MapekPlan>(serviceProvider => {
                 return coordinatorSettings!.UseEuclid ? new EuclidMapekPlan(serviceProvider) : new MapekPlan(serviceProvider);
