+++ models-and-rules/M370-instance.ttl
@@ -16,7 +16,7 @@
                                    "0"^^xsd:int;
               a <http://www.w3.org/ns/sosa/Actuator>,
                 owl:NamedIndividual.
-:EnergyConsumption rdt:hasValue "72.7695803"^^xsd:double;
+:EnergyConsumption rdt:hasValue "16.018624199999998"^^xsd:double;
                    a <http://www.w3.org/ns/sosa/ObservableProperty>,
                      owl:NamedIndividual.
 :EnergyConsumptionDecrease rdt:affectsPropertyWith rdt:ValueDecrease;
@@ -31,7 +31,7 @@
                           owl:NamedIndividual;
                         <http://www.w3.org/ns/sosa/observes> :EnergyConsumption;
                         <http://www.w3.org/ns/ssn/implements> :EnergyConsumptionMeterProcedure.
-:EnergyConsumptionMeterMeasuredEnergyConsumption rdt:hasValue "72.7695803"^^xsd:double;
+:EnergyConsumptionMeterMeasuredEnergyConsumption rdt:hasValue "16.018624199999998"^^xsd:double;
                                                  a <http://www.w3.org/ns/ssn/Property>,
                                                    <http://www.w3.org/ns/ssn/Output>,
                                                    owl:NamedIndividual.
@@ -55,7 +55,7 @@
                   owl:NamedIndividual;
                 <http://www.w3.org/ns/sosa/observes> :RoomHumidity;
                 <http://www.w3.org/ns/ssn/implements> :HumiditySensorProcedure.
-:HumiditySensorMeasuredRoomHumidity rdt:hasValue "36.78305970618142"^^xsd:double;
+:HumiditySensorMeasuredRoomHumidity rdt:hasValue "44.56957331188801"^^xsd:double;
                                     a <http://www.w3.org/ns/ssn/Property>,
                                       <http://www.w3.org/ns/ssn/Output>,
                                       owl:NamedIndividual.
@@ -127,7 +127,7 @@
 :OfficeTemperatureUpperLimit rdt:hasValue "22.0"^^xsd:double;
                              a <http://www.w3.org/ns/ssn/Property>,
                                owl:NamedIndividual.
-:RoomHumidity rdt:hasValue "36.78305970618142"^^xsd:double;
+:RoomHumidity rdt:hasValue "44.56957331188801"^^xsd:double;
               a <http://www.w3.org/ns/sosa/ObservableProperty>,
                 owl:NamedIndividual.
 :RoomHumidityDecrease rdt:affectsPropertyWith rdt:ValueDecrease;
@@ -150,7 +150,7 @@
              rdt:hasURI "roomM370.fmu"^^xsd:anyURI;
              a rdt:FmuModel,
                owl:NamedIndividual.
-:RoomTemperature rdt:hasValue "18.755385596528363"^^xsd:double;
+:RoomTemperature rdt:hasValue "18.565955986928905"^^xsd:double;
                  a <http://www.w3.org/ns/sosa/ObservableProperty>,
                    owl:NamedIndividual.
 :RoomTemperatureIncrease rdt:affectsPropertyWith rdt:ValueIncrease;
@@ -167,7 +167,7 @@
                       owl:NamedIndividual;
                     <http://www.w3.org/ns/sosa/observes> :RoomTemperature;
                     <http://www.w3.org/ns/ssn/implements> :TemperatureSensor1Procedure.
-:TemperatureSensor1MeasuredRoomTemperature rdt:hasValue "18.755385596528363"^^xsd:double;
+:TemperatureSensor1MeasuredRoomTemperature rdt:hasValue "18.565955986928905"^^xsd:double;
                                            a <http://www.w3.org/ns/ssn/Property>,
                                              <http://www.w3.org/ns/ssn/Output>,
                                              <http://www.w3.org/ns/ssn/Input>,
@@ -179,7 +179,7 @@
                       owl:NamedIndividual;
                     <http://www.w3.org/ns/sosa/observes> :RoomTemperature;
                     <http://www.w3.org/ns/ssn/implements> :TemperatureSensor2Procedure.
-:TemperatureSensor2MeasuredRoomTemperature rdt:hasValue "18.755385596528363"^^xsd:double;
+:TemperatureSensor2MeasuredRoomTemperature rdt:hasValue "18.565955986928905"^^xsd:double;
                                            a <http://www.w3.org/ns/ssn/Property>,
                                              <http://www.w3.org/ns/ssn/Output>,
                                              <http://www.w3.org/ns/ssn/Input>,
